Got Albums populated. Working select_exercise.sql. Almost done.

+++ albums_migration.sql
@@ -6,8 +6,7 @@ CREATE TABLE albums (
 	artist VARCHAR(100) NOT NULL,
 	name VARCHAR(100) NOT NULL,
 	release_d YEAR(4) NOT NULL,
-	sales_in_millions FLOAT(15,2) NOT NULL,
+	sales_in_millions FLOAT(8,2) NOT NULL,
 	genre VARCHAR(70) NOT NULL,
 	PRIMARY KEY (id)
 );
-
